+++ activesupport/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Fix `slice!` deleting the default value of the hash.
+
+    *Antonio Santos*
+
 *   `require_dependency` accepts objects that respond to `to_path`, in
     particular `Pathname` instances.